Extracting android defaultConfig block into its own class
